+++ app/src/org/koreader/launcher/device/DeviceInfo.kt
@@ -4,7 +4,7 @@
  * Note: devices don't need to be declared here unless
  * they have known e-ink update routines and/or bug workarounds. */
 
-package org.koreader.launcher
+package org.koreader.launcher.device
 
 import android.os.Build
 import java.util.*
