+++ app/src/org/koreader/launcher/interfaces/EPDInterface.kt
@@ -1,9 +1,9 @@
 /* generic EPD Controller for Android devices,
  * based on https://github.com/unwmun/refreshU */
 
-package org.koreader.launcher
+package org.koreader.launcher.interfaces
 
-interface EPDController {
+interface EPDInterface {
     fun setEpdMode(targetView: android.view.View,
                    mode: Int, delay: Long,
                    x: Int, y: Int, width: Int, height: Int, epdMode: String?)
